+++ lib/macos_ui.dart
@@ -12,7 +12,7 @@ export 'package:flutter/material.dart'
         FlutterLogo,
         CircleAvatar;
 /// todo: package-level docs
-export 'package:flutter/widgets.dart' hide Icon, IconTheme, TextBox;
+export 'package:flutter/widgets.dart' hide Icon, TextBox;
 
 export 'src/buttons/checkbox.dart';
 export 'src/buttons/disclosure_control.dart';
@@ -26,15 +26,22 @@ export 'src/buttons/radio_button.dart';
 export 'src/buttons/switch.dart';
 export 'src/buttons/switch.dart';
 export 'src/buttons/switch.dart';
+export 'src/buttons/switch.dart';
 export 'src/fields_and_labels/combo_box.dart';
 export 'src/fields_and_labels/label.dart';
 export 'src/fields_and_labels/search_field.dart';
 export 'src/fields_and_labels/text_field.dart';
 export 'src/fields_and_labels/token_field.dart';
 export 'src/indicators/progress_indicators.dart';
+export 'src/indicators/progress_indicators.dart';
+export 'src/layout/scaffold.dart';
 export 'src/layout/scaffold.dart';
 export 'src/layout/scaffold.dart';
 export 'src/macos_app.dart';
-export 'src/styles/theme.dart';
+export 'src/macos_app.dart';
+export 'src/styles/macos_theme.dart';
+export 'src/styles/macos_theme_data.dart';
 export 'src/styles/typography.dart';
+export 'src/styles/typography.dart';
+export 'src/util.dart';
 export 'src/util.dart';