+++ lib/src/styles/macos_theme.dart
@@ -1,0 +1,78 @@
+import 'package:flutter/foundation.dart';
+
+import '../../macos_ui.dart';
+import 'macos_theme_data.dart';
+
+/// Applies a macOS-style theme to descendant macOS widgets.
+///
+/// Affects the color and text styles of macOS widgets whose styling
+/// are not overridden when constructing the respective widgets instances.
+///
+/// Descendant widgets can retrieve the current [MacosThemeData] by calling
+/// [MacosThemeData.of]. An [InheritedWidget] dependency is created when
+/// an ancestor [MacosThemeData] is retrieved via [MacosThemeData.of].
+///
+/// See also:
+///
+///  * [MacosThemeData], specifies the theme's visual styling
+///  * [MacosApp], which will automatically add a [MacosTheme] based on the
+///    value of [MacosApp.theme].
+class MacosTheme extends StatelessWidget {
+  /// Creates a [MacosTheme] to change descendant macOS widgets' styling.
+  ///
+  /// The [data] and [child] parameters must not be null.
+  const MacosTheme({
+    Key? key,
+    required this.data,
+    required this.child,
+  }) : super(key: key);
+
+  /// The [MacosThemeData] styling for this theme.
+  final MacosThemeData data;
+
+  /// The widget below this widget in the tree.
+  ///
+  /// {@macro flutter.widgets.ProxyWidget.child}
+  final Widget child;
+
+  static MacosThemeData of(BuildContext context) {
+    final _InheritedMacosTheme? inheritedTheme =
+        context.dependOnInheritedWidgetOfExactType<_InheritedMacosTheme>();
+    return (inheritedTheme?.theme.data ?? MacosThemeData())
+        .resolveFrom(context);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return _InheritedMacosTheme(
+      theme: this,
+      child: child,
+    );
+  }
+
+  @override
+  void debugFillProperties(DiagnosticPropertiesBuilder properties) {
+    super.debugFillProperties(properties);
+    properties.add(
+      DiagnosticsProperty<MacosThemeData>('data', data, showName: false),
+    );
+  }
+}
+
+class _InheritedMacosTheme extends InheritedWidget {
+  const _InheritedMacosTheme({
+    Key? key,
+    required this.theme,
+    required Widget child,
+  }) : super(key: key, child: child);
+
+  final MacosTheme theme;
+
+  @override
+  bool updateShouldNotify(_InheritedMacosTheme old) =>
+      theme.data != old.theme.data;
+}
+
+extension themeContext on BuildContext {
+  MacosThemeData get macosTheme => MacosTheme.of(this);
+}